Refactor tool as base class.
