+++ src/features/emails/components/EmailEditor/tools/InlineLink/index.tsx
@@ -1,82 +1,25 @@
 import getAnchorTags from './utils/getAnchorTags';
+import InlineToolBase from '../../utils/InlineToolBase';
 import {
   API,
   InlineTool,
   InlineToolConstructorOptions,
 } from '@editorjs/editorjs';
 
-let nextId = 1;
-
-export default class LinkTool implements InlineTool {
+export default class LinkTool extends InlineToolBase implements InlineTool {
   private _api: API;
   private _button: HTMLButtonElement | null;
-  private _id: number;
   private _input: HTMLInputElement | null;
-  private _isLink: boolean;
-
-  addLink(range: Range) {
-    const selectedText = range.extractContents();
-    const anchor = document.createElement('a');
-    anchor.classList.add('inlineLink');
-
-    anchor.appendChild(selectedText);
-    range.insertNode(anchor);
-
-    this._api.selection.expandToTag(anchor);
-  }
-
-  checkState() {
-    const range = window.getSelection()?.getRangeAt(0);
-
-    const anchors = range
-      ? getAnchorTags(range, this._api.selection.findParentTag)
-      : [];
-
-    this._isLink = anchors.length > 0;
-
-    if (this._isLink) {
-      this.showInput();
-      return true;
-    } else {
-      this.hideInput();
-      return false;
-    }
-  }
-
-  clear() {
-    this.destroy();
-  }
 
   constructor({ api }: InlineToolConstructorOptions) {
+    super();
     this._api = api;
     this._button = null;
     this._input = null;
-    this._isLink = false;
-    this._id = nextId++;
-    this.handleSelectionChangeBound = () => undefined;
-  }
-
-  destroy() {
-    document.removeEventListener(
-      'selectionchange',
-      this.handleSelectionChangeBound
-    );
-  }
-
-  private handleSelectionChangeBound: () => void | undefined;
-
-  hideInput() {
-    if (this._input) {
-      this._input.hidden = true;
-    }
-  }
-
-  static get isInline() {
-    return true;
   }
 
   removeLink(range: Range) {
-    const anchors = getAnchorTags(range, this._api.selection.findParentTag);
+    const anchors = getAnchorTags(range, this._api);
 
     anchors.forEach((anchor) => {
       anchor.childNodes.forEach((child) =>
@@ -84,28 +27,6 @@ export default class LinkTool implements InlineTool {
       );
       anchor.parentNode?.removeChild(anchor);
     });
-  }
-
-  render() {
-    this._button = document.createElement('button');
-    this._button.type = 'button';
-    this._button.textContent = 'Link';
-    this._button.classList.add(this._api.styles.inlineToolButton);
-
-    const handleSelectionChange = () => {
-      this.checkState();
-
-      if (this._input) {
-        this._input.hidden = !this._isLink;
-      }
-    };
-    this.handleSelectionChangeBound = handleSelectionChange.bind(this);
-    document.addEventListener(
-      'selectionchange',
-      this.handleSelectionChangeBound
-    );
-
-    return this._button;
   }
 
   renderActions() {
@@ -116,6 +37,14 @@ export default class LinkTool implements InlineTool {
     return this._input;
   }
 
+  renderButton() {
+    this._button = document.createElement('button');
+    this._button.type = 'button';
+    this._button.textContent = 'Link';
+    this._button.classList.add(this._api.styles.inlineToolButton);
+    return this._button;
+  }
+
   static get sanitize() {
     return {
       a: {
@@ -124,36 +53,11 @@ export default class LinkTool implements InlineTool {
     };
   }
 
-  showInput() {
-    if (this._input) {
-      this._input.hidden = false;
-    }
+  surround() {
+    // TODO: Either remove link(s) or add link, depending on selection
   }
 
-  get state() {
-    return this._isLink;
-  }
-
-  set state(isLink: boolean) {
-    this._isLink = isLink;
-
-    this._button?.classList.toggle(
-      this._api.styles.inlineToolButtonActive,
-      isLink
-    );
-  }
-
-  surround(range: Range) {
-    if (this._isLink) {
-      this.removeLink(range);
-    } else {
-      this.addLink(range);
-    }
-  }
-
-  static toolbox() {
-    return {
-      title: 'Link',
-    };
+  update() {
+    // TODO: Toggle button and actions
   }
 }
