+++ src/features/emails/components/EmailEditor/tools/InlineLink/utils/getAnchorTags.ts
@@ -1,12 +1,11 @@
-export default function getAnchorTags(
-  range: Range,
-  findParentTag: (tagName: string) => HTMLElement | null
-): Node[] {
+import { API } from '@editorjs/editorjs';
+
+export default function getAnchorTags(range: Range, api: API): Node[] {
   let node: Node | null = range.startContainer;
   const endNode = range.endContainer;
   const anchors: Node[] = [];
 
-  const parentA = findParentTag('A');
+  const parentA = api.selection.findParentTag('A');
 
   if (parentA) {
     anchors.push(parentA);
